+++ lib/segment/src/fixtures/mod.rs
@@ -1,1 +1,2 @@
 pub mod index_fixtures;
+pub mod payload_fixtures;
